// Kniffel-Game/Kniffel/Kniffel-app/src/enums/Category.ts
export enum Category {
    Ones = "Ones",
    Twos = "Twos",
    Threes = "Threes",
    Fours = "Fours",
    Fives = "Fives",
    Sixes = "Sixes",
    ThreeOfAKind = "Three of a Kind",
    FourOfAKind = "Four of a Kind",
    FullHouse = "Full House",
    SmallStreet = "Small Street",
    LargeStreet = "Large Street",
    Kniffel = "Kniffel",
    Chance = "Chance"
}